implement HasCompact for H160
However H160 is encoded/decoded using base fixed length encoding. This
is intentionally 'cause H160 contains some hash and in most cases the
hash will not start with leading zeros. The latter means that in compact
encoding (if it'd be used) H160 uses 21 bytes. This is comparable to
using fixed length encoding.